+++ src/lib/data/openables.ts
@@ -1,9 +1,9 @@
 import LootTable from 'oldschooljs/dist/structures/LootTable';
 
-import { Emoji } from './constants';
-import BirthdayPresentTable from './simulation/birthdayPresent';
-import CasketTable from './simulation/casket';
-import CrystalChestTable from './simulation/crystalChest';
+import { Emoji } from '../constants';
+import BirthdayPresentTable from '../simulation/birthdayPresent';
+import CasketTable from '../simulation/casket';
+import CrystalChestTable from '../simulation/crystalChest';
 
 interface Openable {
 	name: string;
